Don't cram everythin in `lib.rs`
